pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pebblify"
version = "1.0.0"
description = "A comprehensive web scraping and vector database package"
readme = "README.md"
requires-python = ">=3.10,<3.16"
dependencies = [
    "openai>=1.0.0",
    "qdrant-client>=1.7.0",
    "llama-index>=0.9.0",
    "llama-index-vector-stores-qdrant>=0.2.0",
    "llama-index-vector-stores-postgres>=0.2.0",
    "llama-index-embeddings-openai>=0.1.0",
    "llama-index-llms-openai>=0.1.0",
    "numpy>=1.24.0",
    "asyncio-throttle>=1.0.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "crawl4ai",
    "ruff>=0.12.10",
    "fastmcp",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "sentence-transformers>=5.1.0",
    "cohere>=5.17.0",
    "llama-index-embeddings-huggingface>=0.6.0",
    "loguru>=0.7.3",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-xdist>=3.6.1",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.8",
    "types-pyyaml>=6.0.12.20250822",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "ruff>=0.5.0",
    "bandit>=1.7.0",
    "pydocstyle>=6.1.1",
    "detect-secrets>=1.5.0",
    "pre-commit>=3.0.0",
]
[tool.mypy]
ignore_missing_imports = true
warn_unused_configs = true
files = ["core"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = ["__pycache__", "build", "dist", ".venv"]

[tool.ruff.lint]
ignore = ["E203"]

[tool.pydocstyle]
convention = "numpy"
add-ignore = [
    "D100", # allow missing module docstrings in small modules
    "D101", # allow missing class docstrings (will add incrementally)
    "D102", # allow missing method docstrings
    "D103", # allow missing function docstrings
    "D105", # allow magic method docstrings
    "D200", # one-line docstring too long
    "D401", # non-imperative mood
    "D204", # blank line after docstring
    "D406", # Section name punctuation
    "D407", # Section underline
    "D104", # package docstring
]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250822",
]
